fix: code splitting for delete modal
