+++ web/core/components/workspace/index.ts
@@ -3,6 +3,5 @@ export * from "./sidebar";
 export * from "./views";
 export * from "./confirm-workspace-member-remove";
 export * from "./create-workspace-form";
-export * from "./delete-workspace-modal";
 export * from "./logo";
 export * from "./send-workspace-invitation-modal";
